+++ packages/nana-ui/src/settings/index.ts
@@ -1,2 +1,4 @@
 export { default as NanaSettingsPage, default as SettingsPage } from "./NanaSettingsPage.vue";
+export { default as NanaAppearanceSection, default as AppearanceSection } from "./NanaAppearanceSection.vue";
+export { default as NanaAboutSection, default as AboutSection } from "./NanaAboutSection.vue";
 export * from "@lilia/ui-foundation/settings";