+++ packages/nana-ui/src/shell/index.ts
@@ -1,2 +1,3 @@
 export { default as NanaAppShell, default as NanaDesktopShell } from "./NanaAppShell.vue";
 export { default as NanaSidebar } from "./NanaSidebar.vue";
+export { default as NanaTitleBar } from "./NanaTitleBar.vue";