+++ packages/ui-foundation/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./focus";
+export * from "./agentDebug/index";
 export * from "./commands";
 export * from "./settings";
 export * from "./surface";
@@ -12,4 +13,6 @@ export * from "./useRovingFocus";
 export * from "./useSelectionPrimitive";
 export * from "./useSidebarPrimitive";
 export * from "./useThemeSync";
+export * from "./nativeAppearance";
+export * from "./useTauriWindowControls";
 export * from "./useTooltipPrimitive";